refactor: remove compound pattern component
