+++ frontend/src/components/MyApplicationItem/MissionDetail/MissionDetail.module.css
@@ -1,0 +1,15 @@
+.detail-container {
+  display: flex;
+  flex-direction: column;
+  gap: 0.4rem;
+  font-size: 0.875rem;
+  word-break: break-all;
+}
+
+.guide-container {
+  display: flex;
+  gap: 0.25rem;
+  margin-left: auto;
+  color: var(--gray-007);
+  font-size: 0.815rem;
+}
