+++ frontend/src/components/MyApplicationItem/MyApplicationButtons/ApplicationButtons.module.css
@@ -1,11 +1,3 @@
-.button-container {
-  position: absolute;
-  top: 1.25rem;
-  right: 1.5rem;
-  display: flex;
-  gap: 1rem;
-}
-
 .refresh-button {
   background-color: var(--blue-gray);
 }
@@ -19,12 +11,6 @@
 }
 
 @media screen and (max-width: 800px) {
-  .button-container {
-    position: static;
-    gap: 0.4rem;
-    margin-left: auto;
-  }
-
   .judgment-button,
   .apply-button,
   .refresh-button {
