+++ frontend/src/components/MyApplicationItem/MyApplicationItem.module.css
@@ -1,8 +1,4 @@
-.mission-recruit-item {
-  margin-bottom: 1.25rem;
-}
-
-.content-wrapper {
+.content-box {
   position: relative;
   display: flex;
   justify-content: space-between;
@@ -10,6 +6,7 @@
   width: 100%;
   max-width: var(--pc-main-width);
   padding: 1.25rem 1.5rem 1rem;
+  margin-bottom: 1.25rem;
   background-color: var(--white);
   border: 1px solid var(--gray-003);
 }
@@ -19,4 +16,24 @@
   flex-direction: column;
   gap: 0.875rem;
   width: 100%;
+}
+
+.auto-judgment-detail-contour {
+  border: 0.5px solid var(--gray-003);
+}
+
+.button-container {
+  position: absolute;
+  top: 1.25rem;
+  right: 1.5rem;
+  display: flex;
+  gap: 1rem;
+}
+
+@media screen and (max-width: 800px) {
+  .button-container {
+    position: static;
+    gap: 0.4rem;
+    margin-left: auto;
+  }
 }
