+++ frontend/src/components/MyApplicationItem/RecruitmentDetail/RecruitmentDetail.module.css
