--- frontend/src/components/MyApplicationItem/compounds/MissionDetail/MissionDetail.module.css
@@ -1,48 +1,0 @@
-.detail-container {
-  display: flex;
-  flex-direction: column;
-  gap: 0.4rem;
-  font-size: 0.875rem;
-  word-break: break-all;
-}
-
-.pending-container {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.3rem;
-}
-
-.test-result-text {
-  margin-bottom: 0.2rem;
-  font-size: 1.125rem;
-  font-weight: bold;
-}
-
-.test-result-score.pending {
-  color: var(--blue);
-}
-
-.test-result-score.pass {
-  color: var(--green);
-}
-
-.test-result-score.fail {
-  color: var(--red);
-}
-
-.guide-container {
-  display: flex;
-  gap: 0.25rem;
-  margin-left: auto;
-  color: var(--gray-007);
-  font-size: 0.815rem;
-}
-
-.auto-judgment-detail-contour {
-  border: 0.5px solid var(--gray-003);
-}
-
-.commit-hash {
-  color: var(--blue);
-  text-decoration: underline;
-}
